+++ spec/unit/puppet/provider/chocolateysource/windows_spec.rb
@@ -60,6 +60,8 @@ describe provider do
   let (:last_unsupported_version_priority) {'0.9.9.8'}
   let (:minimum_supported_version_bypass_proxy) {'0.10.4'}
   let (:last_unsupported_version_bypass_proxy) {'0.10.3'}
+  let (:minimum_supported_version_allow_self_service) {'0.10.4'}
+  let (:last_unsupported_version_allow_self_service) {'0.10.3'}
   let (:minimum_supported_version_admin_only) {'0.10.8'}
   let (:last_unsupported_version_admin_only) {'0.10.7'}
   let (:minimum_supported_version) {'0.9.9.0'}
@@ -156,6 +158,15 @@ describe provider do
       end
     end
 
+    context ":allow_self_service" do
+      it "should accept 'true' as a string" do
+        resource[:allow_self_service] = 'true'
+      end
+      it "should accept 'true as a boolean'" do
+        resource[:allow_self_service] = true
+      end
+    end
+
     context ":admin_only" do
       it "should accept 'true' as a string" do
         resource[:admin_only] = 'true'
@@ -219,8 +230,8 @@ describe provider do
     element_user = "thisguy"
     element_password = "super/encrypted+value=="
     element_bypass_proxy = "true"
+    element_allow_self_service = "true"
     element_admin_only = "true"
-
 
     before :each do
       element.add_attributes( { "id"          => element_id,
@@ -230,6 +241,7 @@ describe provider do
                                 "user"        => element_user,
                                 "password"    => element_password,
                                 "bypassProxy" => element_bypass_proxy,
+                                "selfService" => element_allow_self_service,
                                 "adminOnly"   => element_admin_only,
                               } )
     end
@@ -247,6 +259,7 @@ describe provider do
       source[:user].must eq element_user
       source[:ensure].must eq :present
       source[:bypass_proxy].must eq element_bypass_proxy
+      source[:allow_self_service].must eq element_allow_self_service
       source[:admin_only].must eq element_admin_only
     end
 
@@ -256,6 +269,7 @@ describe provider do
       element.delete_attribute('user')
       element.delete_attribute('password')
       element.delete_attribute('bypassProxy')
+      element.delete_attribute('selfService')
       element.delete_attribute('adminOnly')
 
       source = provider.get_source(element)
@@ -350,6 +364,33 @@ describe provider do
 
       PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(last_unsupported_version_bypass_proxy)
       Puppet.expects(:warning).with("Chocolatey is unable to specify bypassing system proxy for sources when version is less than 0.10.4. The value you set will be ignored.")
+
+      resource.provider.validate
+    end
+
+    it "should not warn on allow_self_service when choco version is newer than the minimum supported version" do
+      resource[:allow_self_service] = true
+
+      PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(newer_choco_version)
+      Puppet.expects(:warning).never
+
+      resource.provider.validate
+    end
+
+    it "should not warn on allow_self_service when choco version is the minimum supported version" do
+      resource[:allow_self_service] = true
+
+      PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(minimum_supported_version_allow_self_service)
+      Puppet.expects(:warning).never
+
+      resource.provider.validate
+    end
+
+    it "should warn on allow_self_service when choco version is less than the minimum supported version" do
+      resource[:allow_self_service] = true
+
+      PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(last_unsupported_version_allow_self_service)
+      Puppet.expects(:warning).with("Chocolatey is unable to specify self-service for sources when version is less than 0.10.4. The value you set will be ignored.")
 
       resource.provider.validate
     end
@@ -459,6 +500,7 @@ describe provider do
     resource_user = "thatguy"
     resource_password = "secrets!"
     resource_bypass_proxy = :true
+    resource_allow_self_service = :true
     resource_admin_only = :true
 
     before :each do
@@ -494,6 +536,7 @@ describe provider do
       resource[:user] = resource_user
       resource[:password] = resource_password
       resource[:bypass_proxy] = resource_bypass_proxy
+      resource[:allow_self_service] = resource_allow_self_service
       resource[:admin_only] = resource_admin_only
 
       PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(newer_choco_version)
@@ -504,6 +547,7 @@ describe provider do
                                                       '--user', resource_user,
                                                       '--password', resource_password,
                                                       '--bypass-proxy',
+                                                      '--allow-self-service',
                                                       '--admin-only',
                                                       '--priority', resource_priority,
                                                      ])
@@ -543,6 +587,26 @@ describe provider do
                                                       '--name', resource_name,
                                                       '--source', resource_location,
                                                       '--bypass-proxy',
+                                                      '--priority', 0,
+                                                     ])
+
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'enable',
+                                                      '--name', resource_name
+                                                     ])
+
+      resource.flush
+    end
+
+    it "should set allow_self_service when present" do
+      resource[:allow_self_service] = resource_allow_self_service
+
+      PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(newer_choco_version)
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'add',
+                                                      '--name', resource_name,
+                                                      '--source', resource_location,
+                                                      '--allow-self-service',
                                                       '--priority', 0,
                                                      ])
 
@@ -687,6 +751,65 @@ describe provider do
                                                       '--name', resource_name,
                                                       '--source', resource_location,
                                                       '--bypass-proxy',
+                                                      '--priority', 0,
+                                                     ])
+
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'enable',
+                                                      '--name', resource_name,
+                                                     ])
+
+      resource.flush
+    end
+
+    it "should not set allow_self_service when choco version is less than the minimum supported version" do
+      resource[:allow_self_service] = resource_allow_self_service
+
+      PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(last_unsupported_version_allow_self_service)
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'add',
+                                                      '--name', resource_name,
+                                                      '--source', resource_location,
+                                                      '--priority', 0,
+                                                     ])
+
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'enable',
+                                                      '--name', resource_name,
+                                                     ])
+
+      resource.flush
+    end
+
+    it "should set allow_self_service when choco version is newer than the minimum supported version" do
+      resource[:allow_self_service] = resource_allow_self_service
+
+      PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(newer_choco_version)
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'add',
+                                                      '--name', resource_name,
+                                                      '--source', resource_location,
+                                                      '--allow-self-service',
+                                                      '--priority', 0,
+                                                     ])
+
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'enable',
+                                                      '--name', resource_name,
+                                                     ])
+
+      resource.flush
+    end
+
+    it "should set allow_self_service when choco version is the minimum supported version" do
+      resource[:allow_self_service] = resource_allow_self_service
+
+      PuppetX::Chocolatey::ChocolateyCommon.expects(:choco_version).returns(minimum_supported_version_allow_self_service)
+      Puppet::Util::Execution.expects(:execute).with([provider.command(:chocolatey),
+                                                      'source', 'add',
+                                                      '--name', resource_name,
+                                                      '--source', resource_location,
+                                                      '--allow-self-service',
                                                       '--priority', 0,
                                                      ])